% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proj1Fisher.R
\name{fishies}
\alias{fishies}
\title{Fishers Exact Test}
\usage{
fishies(data)
}
\arguments{
\item{data}{Two way table of counts of interest}
}
\value{
The named list of the results of the Fisher's Exact Test.
}
\description{
The null hypothesis of independence leads to the cells
   of the table being distributed according to the hypergeometric distribution.
}
