% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proj1oneway.R
\name{oneway}
\alias{oneway}
\title{One Way Table Analysis}
\usage{
oneway(data, catagories, alpha = 0.05)
}
\arguments{
\item{data}{Vector of counts of interest from the data table}

\item{catagories}{Vector of the names of the categories of the counts}

\item{alpha}{The alpha value to make the desired \eqn{100(1-alpha)\%} CI}
}
\value{
The named list of the results of the chi squared test of uniformity.
}
\description{
This function takes the data from a one way table, makes a plot of the proportions with CIs,
   and runs a chi squared test of uniformity.
}
