% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proj1twoway.R
\name{twoway}
\alias{twoway}
\title{Two Way Table Analysis}
\usage{
twoway(data, nums, cat1, cat2, alpha = 0.1)
}
\arguments{
\item{data}{List of complete 2-way data set with each value being associated with its specific factors.}

\item{nums}{Two way table of counts by both factors}

\item{cat1}{One way table of counts by first factor}

\item{cat2}{One way table of counts by second factor}

\item{alpha}{The alpha value to make the desired \eqn{100(1-alpha)\%} CI}
}
\value{
The named list of the results of the chi squared test of independence.
}
\description{
This function takes the data from a two way table, makes plots of the proportions of themarginal
   distributions of each factor with CIs, makes a plot of the overall distribution with CIs,
   and runs a chi squared test of independence of factors.
}
